+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - added itemStackIngredients to RPGCraftingRecipe. This is a separate list of ingredients that consists of item stacks. When data components are defined for a stack, a stack with the exact same data components is required for the recipe. If no components are defined, any stack of that item type will work. Both types of ingredient lists are active at the same time. Note that existing recipes of the "rpgcrafting:rpg_crafting_recipe" recipe type must be updated!
 - added (in)active tab provider block.
 - potentially fixed lingering crafting result slots
+- hand-crafting screen now shows the hand-crafting level
 
 # 1.0.0